Support engine override, support overriding some env, and improve macOS compatibility
With the latest Docker on Homebrew, build no longer works. It seems an engine
override is a nice addition, in case you prefer to not select docker as the
default choice.

Additionally, to allow forks to be more maintainable, allow the docker tag,
repository, ZBM, and EFI to be changed through env. These are exposed through
env since most users should not customize these.

Use the latest released ZBM as the default.

Improve macOS compatibility by using GNU versions for du and stat.
inject (stat), and iso (du) are fixed by checking for GNU versions.

Signed-off-by: Daniel Vicory <dvicory@gmail.com>